Pull deposit logic into its own controller endpoint and service method

+++ src/main/resources/db/migration/V2__create-bank-account-table.sql
@@ -1,12 +1,12 @@
 CREATE TABLE IF NOT EXISTS bank_accounts
 (
 	id                BIGSERIAL PRIMARY KEY,
-	user_id           BIGINT       NOT NULL REFERENCES users (id),
-	account_type      VARCHAR(50)  NOT NULL,
-	wallet_identifier VARCHAR(255) NOT NULL UNIQUE,
-	classic_address   VARCHAR(255) NOT NULL UNIQUE,
-	x_address         VARCHAR(255) NOT NULL UNIQUE,
-	balance           BIGINT       NOT NULL,
-	created_at        TIMESTAMPTZ  NOT NULL DEFAULT NOW(),
-	updated_at        TIMESTAMPTZ  NOT NULL DEFAULT NOW()
+	user_id           BIGINT         NOT NULL REFERENCES users (id),
+	account_type      VARCHAR(50)    NOT NULL,
+	wallet_identifier VARCHAR(255)   NOT NULL UNIQUE,
+	classic_address   VARCHAR(255)   NOT NULL UNIQUE,
+	x_address         VARCHAR(255)   NOT NULL UNIQUE,
+	balance           NUMERIC(20, 2) NOT NULL,
+	created_at        TIMESTAMPTZ    NOT NULL DEFAULT NOW(),
+	updated_at        TIMESTAMPTZ    NOT NULL DEFAULT NOW()
 )
